django-service/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
gunicorn==21.2.0
python-decouple==3.8
